+++ src/test/java/org/sonar/plugins/cas/AuthenticationFilterTest.java
@@ -2,10 +2,10 @@ package org.sonar.plugins.cas;
 
 import org.junit.Test;
 import org.sonar.api.config.Configuration;
+import org.sonar.api.server.http.HttpRequest;
+import org.sonar.api.server.http.HttpResponse;
+import org.sonar.api.web.FilterChain;
 
-import javax.servlet.FilterChain;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 import static org.mockito.Mockito.mock;
@@ -18,8 +18,8 @@ public class AuthenticationFilterTest {
         Configuration config = new SonarTestConfiguration()
                 .withAttribute("sonar.cas.casServerLoginUrl", "https://cas.server.net/cas/login")
                 .withAttribute("sonar.cas.sonarServerUrl", "https://sonar.server.net");
-        HttpServletRequest request = mock(HttpServletRequest.class);
-        HttpServletResponse response = mock(HttpServletResponse.class);
+        HttpRequest request = mock(HttpRequest.class);
+        HttpResponse response = mock(HttpResponse.class);
 
         FilterChain filterChain = mock(FilterChain.class);