+++ src/test/java/org/sonar/plugins/cas/LoginHandlerTest.java
@@ -1,10 +1,9 @@
 package org.sonar.plugins.cas;
 
 import org.junit.Test;
+import org.sonar.api.server.http.HttpRequest;
 import org.sonar.plugins.cas.session.CasSessionStore;
 import org.sonar.plugins.cas.session.CasSessionStoreFactory;
-
-import javax.servlet.http.HttpServletRequest;
 
 import static org.fest.assertions.Assertions.assertThat;
 import static org.mockito.Mockito.*;
@@ -13,7 +12,7 @@ import static org.sonar.plugins.cas.LoginHandler.getTicketParameter;
 public class LoginHandlerTest {
     @Test
     public void getTicketParameterShouldReturnEmptyString() {
-        HttpServletRequest request = mock(HttpServletRequest.class);
+        HttpRequest request = mock(HttpRequest.class);
         when(request.getParameter("ticket")).thenReturn(null);
 
         CasSessionStore sessionStore = mock(CasSessionStore.class);
@@ -30,7 +29,7 @@ public class LoginHandlerTest {
 
     @Test
     public void getTicketParameterShouldReturnTicketParameter() {
-        HttpServletRequest request = mock(HttpServletRequest.class);
+        HttpRequest request = mock(HttpRequest.class);
         // even when the ticket is called ST-XYZZY it could represent both a proxy ticket or a service ticket
         when(request.getParameter("ticket")).thenReturn("ST-012345678");